pyproject.toml
[tool.poetry]
name = "learnsherpa-ai"
version = "0.1.0"
description = "LearnSherpa_AI using crewAI"
authors = ["Manel Soler <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.51.0" }
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
learn_sherpa_ai = "src.main:run"
run_crew = "src.main:run"
train = "src.main:train"
replay = "src.main:replay"
test = "src.main:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
